Naming and file structure updated, minor clean ups and refactors made
