+++ mithril/framework/physical/model.py
@@ -55,7 +55,7 @@ from ..logical.model import (
     Model,
     define_unique_names,
 )
-from ..logical.primitive import PrimitiveModel
+from ..logical.operator import Operator
 from .data_store import StaticDataStore
 from .flat_graph import FlatGraph
 
@@ -221,7 +221,7 @@ class PhysicalModel(GenericDataType[DataType]):
             # solver to propagate updates.
             self.flat_graph.constraint_solver(updates)
 
-            output = PrimitiveModel.output_key
+            output = Operator.output_key
             _data_dict: dict[str, IOHyperEdge] = {}
 
             self._infer_differentiability(model_data)
@@ -233,7 +233,7 @@ class PhysicalModel(GenericDataType[DataType]):
 
             # NOTE: maybe move adding cache to generate_code methods.
             if self.backend.backend_type == "numpy":
-                cache_name = "_".join([mappings[output], PrimitiveModel.cache_name])
+                cache_name = "_".join([mappings[output], Operator.cache_name])
                 mappings["cache"] = cache_name
                 # TODO: Why do we have to provide cache_value here? It is
                 # NONE | dict().
@@ -417,7 +417,7 @@ class PhysicalModel(GenericDataType[DataType]):
     def _infer_differentiability(self, model_data: dict[str, IOHyperEdge]) -> None:
         # Infer output differentiability only for the models
         # that have a Tensor type output.
-        output_key = PrimitiveModel.output_key
+        output_key = Operator.output_key
         output_edge = model_data[output_key]
         if output_edge.is_tensor:
             # If any of the inputs are differentiable, then
@@ -556,9 +556,7 @@ class PhysicalModel(GenericDataType[DataType]):
                     )
                 )
             ):
-                self.ignore_grad_keys.add(
-                    node.connections[PrimitiveModel.output_key].key
-                )
+                self.ignore_grad_keys.add(node.connections[Operator.output_key].key)
 
         if len(self._output_keys - self.ignore_grad_keys) == 0 and not self.inference:
             raise ValueError("All outputs gradient are ignored.")
@@ -800,7 +798,7 @@ class PhysicalModel(GenericDataType[DataType]):
             table.display()
             if depth > 0:
                 for model, model_name in name_mappings.items():
-                    if not isinstance(model, PrimitiveModel):
+                    if not isinstance(model, Operator):
                         self.summary(
                             model=model,
                             depth=depth - 1,
@@ -813,7 +811,7 @@ class PhysicalModel(GenericDataType[DataType]):
                         )
 
     def extract_connection_info(
-        self, name_mappings: dict[PrimitiveModel, str] | None = None
+        self, name_mappings: dict[Operator, str] | None = None
     ) -> dict[str, tuple[dict[str, list[str]], dict[str, list[str]]]]:
         if name_mappings is None:
             name_mappings = define_unique_names(self.flat_graph.get_models())  # type: ignore
@@ -885,7 +883,7 @@ class PhysicalModel(GenericDataType[DataType]):
 
     def _replace_with_primitive(
         self, model: Model, key_mappings: dict[str, str]
-    ) -> tuple[PrimitiveModel, dict[str, str]]:
+    ) -> tuple[Operator, dict[str, str]]:
         assert model.formula_key is not None
         formula = self.backend.primitive_function_dict[model.formula_key]
         primitive_input_keys = formula.__code__.co_varnames[
@@ -915,9 +913,7 @@ class PhysicalModel(GenericDataType[DataType]):
 
         kwargs = {key: model.conns.all[key].metadata for key in external_keys}
 
-        primitive = PrimitiveModel(
-            formula_key=model.formula_key, name=model.name, **kwargs
-        )
+        primitive = Operator(formula_key=model.formula_key, name=model.name, **kwargs)
         primitive.parent = model.parent
 
         p_key_mappings: dict[str, str] = {}
@@ -1026,7 +1022,7 @@ class FlatModel:
             short_namings (bool): Flag to determine if short namings should be used.
         """
 
-        self.mappings: dict[PrimitiveModel, dict[str, Name]] = {}
+        self.mappings: dict[Operator, dict[str, Name]] = {}
         self.assigned_edges: dict[IOHyperEdge, Name] = {}
         self.assigned_names: dict[str, Name] = {}
         self.external_edges: dict[IOHyperEdge, str] = {}
@@ -1034,7 +1030,7 @@ class FlatModel:
         self.key_origins: dict[str, int] = {}
         self.reserved_keys: set[str] = reserved_keys if reserved_keys else set()
         self.queued_models: dict[
-            IOHyperEdge, list[tuple[PrimitiveModel, dict[str, str], str]]
+            IOHyperEdge, list[tuple[Operator, dict[str, str], str]]
         ] = {}
         self._external_mapping: dict[str, Name] = {}
         self.model = model
@@ -1188,7 +1184,7 @@ class FlatModel:
         if mappings is None:
             mappings = {}
 
-        if isinstance(model, PrimitiveModel):
+        if isinstance(model, Operator):
             if not self._is_primitive_ready(model):
                 self._add_primitive_to_queue(model, mappings, parent_name)
                 return
@@ -1198,16 +1194,16 @@ class FlatModel:
         elif isinstance(model, Model):
             self._process_model(model, mappings, parent_name)
         else:
-            raise ValueError("Model must be either PrimitiveModel or Model")
+            raise ValueError("Model must be either Operator or Model")
 
     def _process_primitive_model(
-        self, model: PrimitiveModel, mappings: dict[str, str], parent_name: str
+        self, model: Operator, mappings: dict[str, str], parent_name: str
     ) -> None:
         """
         Process a primitive model.
 
         Args:
-            model (PrimitiveModel): The primitive model.
+            model (Operator): The primitive model.
             mappings (dict[str, str]): The mappings of keys.
         """
 
@@ -1280,12 +1276,12 @@ class FlatModel:
                         m, mappings=mappings, parent_name=parent_name
                     )
 
-    def _is_primitive_ready(self, model: PrimitiveModel) -> bool:
+    def _is_primitive_ready(self, model: Operator) -> bool:
         """
         Check if a primitive model is ready to be processed.
 
         Args:
-            model (PrimitiveModel): The primitive model.
+            model (Operator): The primitive model.
 
         Returns:
             bool: True if the model is ready, False otherwise.
@@ -1297,13 +1293,13 @@ class FlatModel:
         return True
 
     def _add_primitive_to_queue(
-        self, model: PrimitiveModel, mappings: dict[str, str], parent_name: str
+        self, model: Operator, mappings: dict[str, str], parent_name: str
     ) -> None:
         """
         Add a primitive model to the queue.
 
         Args:
-            model (PrimitiveModel): The primitive model.
+            model (Operator): The primitive model.
             input_edges (set[IOHyperEdge]): The input edges.
             mappings (dict[str, str]): The mappings of keys.
         """
@@ -1365,6 +1361,6 @@ class FlatModel:
         self._iter = iter(self.mappings.items())
         return self
 
-    def __next__(self) -> tuple[PrimitiveModel, dict[str, str]]:
+    def __next__(self) -> tuple[Operator, dict[str, str]]:
         model, mapping = next(self._iter)
         return model, {key: name.name for key, name in mapping.items()}
