+++ mithril/models/models.py
@@ -35,68 +35,66 @@ from ..framework.logical.model import (
     IOKey,
     Model,
 )
-from ..framework.logical.user_essential_primitives import (
+from ..utils.utils import PaddingType, convert_to_list, convert_to_tuple
+from .primitives import (
     Absolute,
     Add,
     ArgMax,
-    Buffer,
-    Cast,
-    Divide,
-    Dtype,
-    Exponential,
-    Greater,
-    Indexer,
-    Length,
-    MatrixMultiply,
-    Mean,
-    Minus,
-    Multiply,
-    Power,
-    Reshape,
-    Shape,
-    Size,
-    Slice,
-    Sqrt,
-    Subtract,
-    Sum,
-    Transpose,
-    Variance,
-)
-from ..utils.utils import PaddingType, convert_to_list, convert_to_tuple
-from .primitives import (
     AUCCore,
+    Buffer,
     CartesianDifference,
+    Cast,
     Cholesky,
     Concat,
     DistanceMatrix,
+    Divide,
+    Dtype,
     Eigvalsh,
+    Exponential,
     Eye,
     EyeComplement,
     GPRAlpha,
     GPRVOuter,
+    Greater,
+    Indexer,
     KLDivergence,
+    Length,
     Log,
+    MatrixMultiply,
+    Mean,
+    Minus,
+    Multiply,
     NormModifier,
     PaddingConverter1D,
     PaddingConverter2D,
     PermuteTensor,
     PolynomialFeatures,
+    Power,
     PrimitiveConvolution1D,
     PrimitiveConvolution2D,
     PrimitiveMaxPool1D,
     PrimitiveMaxPool2D,
+    Reshape,
+    Shape,
     Sigmoid,
     Sign,
+    Size,
+    Slice,
+    Sqrt,
     Square,
     Squeeze,
     StableReciprocal,
     StrideConverter,
+    Subtract,
+    Sum,
     Tanh,
+    Transpose,
     TransposedDiagonal,
     Trapezoid,
     TsnePJoint,
     TupleConverter,
     Unique,
+    Variance,
     Where,
 )
