+++ tests/scripts/test_codegen.py
@@ -423,6 +423,9 @@ def test_default_kwarg_reduction_2(file_path: str):
         compare_callables(evaluate, eval_func)
 
 
+a = 1
+
+
 @with_temp_file(".py")
 def test_array_creation_primitive(file_path: str):
     model = Model()
